Refactoring objects and endpoints.
