+++ src/Presentation/CleanArchitectureSampleProject.Presentation.MinimalAPI/Endpoints/Products/ProductsEndpoints.GetByName.cs
@@ -1,0 +1,42 @@
+﻿using CleanArchitectureSampleProject.Application.Outputs;
+using CleanArchitectureSampleProject.Application.UseCases;
+using CleanArchitectureSampleProject.CrossCuttingConcerns;
+using Microsoft.AspNetCore.Mvc;
+using System.Net;
+
+namespace CleanArchitectureSampleProject.Presentation.MinimalAPI.Endpoints.Products;
+
+public static partial class ProductsEndpoints
+{
+    private static WebApplication MapGetByName(this WebApplication app)
+    {
+        app.MapGet($"/{Controller}/by-name/{{productName}}", async (ILogger<Logging> logger, IProductUseCases productUseCases, string productName, CancellationToken cancellation) =>
+        {
+            return await GetByName(logger, productUseCases, productName, cancellation);
+        })
+        .Produces<GetProductOutput>(Success, ContentType)
+        .Produces<ProblemDetails>(BadRequest, ContentType)
+        .WithConfigSummaryInfo("Get Product By Name", TagName);
+        return app;
+    }
+
+    private static async Task<IResult> GetByName(ILogger<Logging> logger, IProductUseCases productUseCases, string productName, CancellationToken cancellation)
+    {
+        const string errorTitle = "Error while getting product by name.";
+
+        var result = await productUseCases.GetProductByName(productName, cancellation);
+        return result.Match(success => Results.Ok(success),
+            error =>
+            {
+                var errorMessage = logger.LogSeqError(error);
+                return Results.Problem(
+                    type: HttpStatusCode.BadRequest.ToString(),
+                    title: errorTitle,
+                    detail: errorMessage,
+                    statusCode: StatusCodes.Status400BadRequest
+                );
+            }
+        );
+    }
+
+}
